+++ tonapi/src/main/kotlin/io/tonapi/models/DecodedMessageExtInMsgDecodedWalletV3.kt
@@ -33,13 +33,13 @@ import com.squareup.moshi.JsonClass
 data class DecodedMessageExtInMsgDecodedWalletV3 (
 
     @Json(name = "subwallet_id")
-    val subwalletId: kotlin.Int,
+    val subwalletId: kotlin.Long,
 
     @Json(name = "valid_until")
-    val validUntil: kotlin.Int,
+    val validUntil: kotlin.Long,
 
     @Json(name = "seqno")
-    val seqno: kotlin.Int,
+    val seqno: kotlin.Long,
 
     @Json(name = "raw_messages")
     val rawMessages: kotlin.collections.List<DecodedRawMessage>
